Abstract confession events to BaseConfessionEvent
