+++ app/Events/ConfessionStatusWasChanged.php
@@ -2,22 +2,13 @@
 
 namespace NUSWhispers\Events;
 
-use Illuminate\Queue\SerializesModels;
 use NUSWhispers\Models\Confession;
 use NUSWhispers\Models\User;
 
-class ConfessionStatusWasChanged
+class ConfessionStatusWasChanged extends BaseConfessionEvent
 {
-    use SerializesModels;
-
-    /** @var \NUSWhispers\Models\Confession */
-    public $confession;
-
     /** @var string */
     public $originalStatus;
-
-    /** @var \NUSWhispers\Models\User */
-    public $user;
 
     /**
      * Create a new event instance.
@@ -28,8 +19,7 @@ class ConfessionStatusWasChanged
      */
     public function __construct(Confession $confession, $originalStatus, User $user = null)
     {
-        $this->confession = $confession;
+        parent::__construct($confession, $user);
         $this->originalStatus = $originalStatus;
-        $this->user = $user;
     }
 }
